Remove legacy code for k10cr1 and kbd101
