+++ hardware_tests/test_thorlabs_k10cr1.py
@@ -12,7 +12,7 @@ from pnpq.apt.protocol import (
     LimitSwitch,
     StopMode,
 )
-from pnpq.devices.refactored_waveplate_thorlabs_k10cr1 import WaveplateThorlabsK10CR1
+from pnpq.devices.waveplate_thorlabs_k10cr1 import WaveplateThorlabsK10CR1
 from pnpq.units import pnpq_ureg
 
 log = structlog.get_logger()
