--- tests/test_odl_thorlabs_kbd101.py
@@ -1,8 +1,0 @@
-import pytest
-
-from pnpq.devices.odl_thorlabs_kbd101 import OdlThorlabs
-
-
-def test_disconnected_initialization() -> None:
-    with pytest.raises(RuntimeError):
-        OdlThorlabs("ABC", "DEF")
